+++ wfsim/load_resource.py
@@ -1,10 +1,11 @@
 from copy import deepcopy
-import os.path as osp
-
-import numpy as np
-import strax
-import straxen
 import logging
+import numpy as np
+import os.path as osp
+import strax
+from strax import exporter
+import straxen
+
 
 NT_AUX_INSTALLED = False
 STRAX_AUX_INSTALLED = False
@@ -20,15 +21,17 @@ try:
 except (ModuleNotFoundError, ImportError):
     pass
 
-logging.basicConfig(handlers=[
-    # logging.handlers.WatchedFileHandler('wfsim.log'),
-    logging.StreamHandler()])
+
+export, __all__ = exporter()
+logging.basicConfig(handlers=[logging.StreamHandler()])
 log = logging.getLogger('wfsim.resource')
 log.setLevel('WARNING')
+
 
 _cached_configs = dict()
 
 
+@export
 def load_config(config):
     """Create a Resource instance from the configuration
 
@@ -43,6 +46,7 @@ def load_config(config):
     return result
 
 
+@export
 class Resource:
     """
     Get the configs needed for running WFSim. Configs can be obtained in
@@ -182,6 +186,11 @@ class Resource:
                 continue
             if k == 'url_base':
                 continue
+            if v == '':
+                log.warning(f'{k} has no path so this config file is set to None')
+                files[k] = None
+                continue
+
             log.debug(f'Obtaining {k} from {v}')
             files[k] = self.get_file_path(files['url_base'], v)
 
@@ -250,7 +259,8 @@ class Resource:
 
             # S1 photon timing splines
             if config.get('s1_time_spline', False):
-                self.s1_time_splines = straxen.get_resource(files['s1_time_spline'], fmt='pkl')
+                self.s1_optical_propagation_spline = make_map(files['s1_time_spline'], fmt='json.gz',
+                                                              method='RegularGridInterpolator')
 
             # Electron After Pulses
             if config.get('enable_electron_afterpulses', False):
@@ -271,7 +281,7 @@ class Resource:
         log.debug(f'{self.__class__.__name__} fully initialized')
 
 
-def make_map(map_file, fmt='text'):
+def make_map(map_file, fmt='text', method='WeightedNearestNeighbors'):
     """Fetch and make an instance of InterpolatingMap based on map_file
     Alternatively map_file can be a list of ["constant dummy", constant: int, shape: list]
     return an instance of  DummyMap"""
@@ -284,12 +294,13 @@ def make_map(map_file, fmt='text'):
     elif isinstance(map_file, str):
         log.debug(f'Initialize map interpolator for file {map_file}')
         map_data = straxen.get_resource(map_file, fmt=fmt)
-        return straxen.InterpolatingMap(map_data)
+        return straxen.InterpolatingMap(map_data, method=method)
 
     else:
         raise TypeError("Can't handle map_file except a string or a list")
 
 
+@export
 class DummyMap:
     """Return constant results
         the length match the length of input